DBZ-8924 Add LogMinerQueryBuilder contract
